% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank_data}
\alias{bank_data}
\title{Bank marketing campaigns data set analysis}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{job}{Type of job}
 \item{marital}{marital status}
 \item{education}{education}
 \item{default}{has credit in default?}
 \item{housing}{has housing loan?}
 \item{loan}{has personal loan?}
 \item{age}{age}
 \item{y}{has the client subscribed a term deposit? (binary: "yes","no")}
}
}
\usage{
data(bank)
}
\description{
It is a dataset that describing Portugal bank marketing campaigns results.
Conducted campaigns were based mostly on direct phone calls, offering bank client to place a term deposit.
If after all marking efforts client had agreed to place deposit - target variable marked 'yes', otherwise 'no'
}
\details{
Sourse of the data
https://archive.ics.uci.edu/ml/datasets/bank+marketing
}
\examples{

data(bank)
head(bank_data)
}
\references{
This dataset is public available for research. The details are described in S. Moro, P. Cortez and P. Rita. A Data-Driven Approach to Predict the Success of Bank Telemarketing. Decision Support Systems, Elsevier, 62:22-31, June 2014
}
\keyword{datasets}
