% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut3.R
\name{cut3}
\alias{cut3}
\title{cut3}
\usage{
cut3(
  x,
  cuts,
  m = 150,
  g,
  levels.mean = FALSE,
  digits,
  minmax = TRUE,
  oneval = TRUE,
  onlycuts = FALSE,
  formatfun = format,
  ...
)
}
\arguments{
\item{x}{description}

\item{cuts}{description}

\item{m}{description}

\item{g}{description}

\item{levels.mean}{description}

\item{digits}{description}

\item{minmax}{description}

\item{oneval}{description}

\item{onlycuts}{description}

\item{formatfun}{description}
}
\value{

}
\description{
Copy of `Hmisc::cut2()` to reduce dependency on `survival` package. Function like cut but left endpoints are inclusive and labels are of the form [lower, upper), except that last interval is [lower,upper].
}
\examples{
library(testthat)
library(prettyglm)
data("bank")
columns_to_factor <- c('job',
                       'marital',
                       'education',
                       'default',
                       'housing',
                       'loan')
bank_data  <- bank_data  \%>\%
  dplyr::mutate_at(columns_to_factor, list(~factor(.))) \%>\%
  dplyr::mutate(age_cat = cut3(age, g=30, levels.mean=TRUE))

}
\author{
F. Harrell
}
