% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_outcome.R
\name{predict_outcome}
\alias{predict_outcome}
\title{predict_outcome}
\usage{
predict_outcome(
  target,
  model_object,
  dataset,
  prediction_type = NULL,
  weights = NULL
)
}
\arguments{
\item{target}{String of target variable name.}

\item{model_object}{Model object. prettyglm currently supports}

\item{dataset}{This is used to plot the number in each class as a barchart if plotly is TRUE.}

\item{prediction_type}{type of prediction to be passed to the model object. For ...GLM defaults to ....}

\item{weights}{weightings to be provided to predictions if required.}
}
\value{
\item{dataframe}{Returns a dataframe of Actual and Predicted Values}
}
\description{
Processing to predict response for various actual vs expected plots
}
\seealso{
\code{\link[broom]{tidy.lm}}
}
\author{
Jared Fowler
}
