% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual_expected_bucketed.R
\name{actual_expected_bucketed}
\alias{actual_expected_bucketed}
\title{actual_expected_bucketed}
\usage{
actual_expected_bucketed(
  Target_Variable,
  Model_Object,
  Data_Set,
  number_of_buckets = 25,
  plotlyplot = T,
  mywidth = 800,
  myheight = 500,
  Return_data = F,
  first_colour = "black",
  second_colour = "#f185f2"
)
}
\arguments{
\item{Target_Variable}{String of target variable name.}

\item{Model_Object}{GLM model object.}

\item{Data_Set}{Training data used when building the model. This is used to plot the number in each class as a barchart if plotly is TRUE.}

\item{plotlyplot}{Logical if returned plot should be a plotly with number of records as a second axis.}

\item{mywidth}{Plotly plot width in pixels.}

\item{myheight}{Plotly plot height in pixels.}

\item{Return_data}{Logical to return cleaned dataset instead of plot.}

\item{first_colour}{First colour to plot, usually the colour of actual.}

\item{second_colour}{Second colour to plot, usually the colour of predicted.}
}
\value{
plot
Plotly plot by defualt.
ggplot if plotlyplot = F.
Tibble if Return_data = T.
}
\description{
Provides a rank plot of the actual and predicted.
}
\examples{
\dontrun{
}
}
