% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual_expected_bucketed.R
\name{actual_expected_bucketed}
\alias{actual_expected_bucketed}
\title{actual_expected_bucketed}
\usage{
actual_expected_bucketed(
  target_variable,
  model_object,
  data_set = NULL,
  number_of_buckets = 25,
  width = 800,
  height = 500,
  Return_data = F,
  first_colour = "black",
  second_colour = "#cc4678",
  predict_function = NULL
)
}
\arguments{
\item{target_variable}{String of target variable name.}

\item{model_object}{GLM model object.}

\item{data_set}{Training data used when building the model. This is used to plot the number in each class as a barchart if plotly is TRUE.}

\item{width}{Plotly plot width in pixels.}

\item{height}{Plotly plot height in pixels.}

\item{Return_data}{Logical to return cleaned dataset instead of plot.}

\item{first_colour}{First colour to plot, usually the colour of actual.}

\item{second_colour}{Second colour to plot, usually the colour of predicted.}
}
\value{
plot
Plotly plot by defualt.
ggplot if plotlyplot = F.
Tibble if Return_data = T.
}
\description{
Provides a rank plot of the actual and predicted.
}
\examples{
\dontrun{
}
}
