% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut3.R
\name{cut3}
\alias{cut3}
\title{cut3}
\usage{
cut3(
  x,
  cuts,
  m = 150,
  g,
  digits,
  minmax = TRUE,
  oneval = TRUE,
  onlycuts = FALSE,
  formatfun = format,
  ...
)
}
\arguments{
\item{x}{numeric vector to classify into intervals.}

\item{cuts}{cut points.}

\item{m}{desired minimum number of observations in a group. The algorithm does not guarantee that all groups will have at least m observations.}

\item{g}{number of quantile groups}

\item{digits}{number of significant digits to use in constructing levels.}

\item{minmax}{if cuts is specified but min(x)<min(cuts) or max(x)>max(cuts), augments cuts to include min and max x}

\item{oneval}{if an interval contains only one unique value, the interval will be labeled with the formatted version of that value instead of the interval endpoints, unless oneval=FALSE}

\item{onlycuts}{set to TRUE to only return the vector of computed cuts. This consists of the interior values plus outer ranges.}

\item{formatfun}{format function}

\item{...}{additional arguments passed to formatfun}
}
\value{
vector of cut
}
\description{
Hmisc::cut2 bones repackaged to remove errors with importing Hmisc
}
