% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_coefficients.R
\name{pretty_coefficients}
\alias{pretty_coefficients}
\title{pretty_coefficients}
\usage{
pretty_coefficients(
  model_object,
  relativity_transform = NULL,
  type_iii = NULL,
  conf.int = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{model_object}{Model object to create coefficient table for. Must be of type: \link[stats:glm]{stats::glm()}, \link[stats:lm]{stats::lm()},  \link[parsnip:linear_reg()]{parsnip::linear_reg()}, \link[parsnip:logistic_reg()]{parsnip::logistic_reg()} or  \link[poissonreg:poisson_reg()]{poissonreg::poisson_reg()}}

\item{relativity_transform}{String of the function to be applied to the model estimate to calculate the relativity, for example: 'exp(estimate)-1'. Default is for relativity to be excluded from output.}

\item{type_iii}{Type III statistical test to perform. Default is none. Options are 'Wald' or 'LR'. Warning 'LR' can be computationally expensive. Test performed via \link[car:Anova()]{car::Anova()}}

\item{conf.int}{Set to TRUE to include confidence intervals in summary table. Warning, can be computationally expensive.}

\item{return_data}{Set to TRUE to return \link[base:data.frame]{base::data.frame()} instead of creating \link[knitr:kable]{knitr::kable()}.}
}
\value{
\link[knitr:kable]{knitr::kable()} if return_data = FALSE. \link[base:data.frame]{base::data.frame()} if return_data = TRUE.
}
\description{
Creates a pretty kable of model coefficients including coefficient base levels.
}
\examples{
\dontrun{
}
}
