% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineit.R
\name{splineit}
\alias{splineit}
\title{splineit}
\usage{
splineit(var, min, max)
}
\arguments{
\item{var}{Continous vector to spline.}

\item{min}{Min of spline.}

\item{max}{Max of spline.}
}
\value{
Splined Column
}
\description{
Splines a continous variable
}
\examples{
# add an example here
}
